Create csv2wiki.pl
Just what it says it is, it takes a CSV and turns it into a mediaWiki formatted table. At the moment it takes files as input, could easily be modified to work as a cut and paste. (I may do this later, but again: Do X so I can get back to Y)